requirements.txt
requests==2.31.0
confluent-kafka==2.3.0 
psycopg2-binary==2.9.9